- add команду обработки mortal combat /mortal_combat 114 - fix time формат при команде /time - конфиги для AMBUSH - засада больше не затрагивает игроков ниже 15 уровня - announce предупреждающий о наступлении ночи и предупреждение срабатывающее каждые 4 часа.

+++ src/main/java/net/sf/l2j/gameserver/model/actor/instance/Monster.java
@@ -68,7 +68,7 @@ public class Monster extends Attackable {
     private boolean _isRaid;
     private boolean _isMinion;
     @Getter @Setter
-    private float nightExpSpBonus;
+    private double nightExpSpBonus;
 
     public Monster(int objectId, NpcTemplate template) {
         super(objectId, template);
